requirements.txt
pyserial==3.5
scapy==2.5.0
